+++ lib/calendrical/time.ex
@@ -55,6 +55,8 @@ defmodule Calendrical.Time do
   * `{:error, Calendrical.TimeParseError.t()}` when no pattern
     matched.
 
+  * `{:error, Localize.InvalidValueError.t()}` when `input` is not a string or an option is malformed.
+
   ### Examples
 
       iex> Calendrical.Time.parse("14:30:00", locale: :en)
@@ -75,7 +77,9 @@ defmodule Calendrical.Time do
   """
   @spec parse(String.t(), Keyword.t()) ::
           {:ok, Time.t() | map()} | {:error, Exception.t()}
-  def parse(input, options \\ []) when is_binary(input) do
-    Calendrical.Time.Parser.parse(input, options)
+  def parse(input, options \\ []) do
+    with :ok <- Calendrical.ParseOptions.validate(input, options) do
+      Calendrical.Time.Parser.parse(input, options)
+    end
   end
 end